+++ community/cypher/cypher/src/main/java/org/neo4j/cypher/internal/codegen/CompiledEquivalenceUtils.java
@@ -555,7 +555,7 @@ public final class CompiledEquivalenceUtils
 
     private static boolean mixedFloatEquality( Float a, Double b )
     {
-        return a == b.floatValue();
+        return a.doubleValue() == b;
     }
 }
